ugh okay rcscores should actually be correct now, except maybe not because the search result is screwed up
git-svn-id: svn://svn.code.sf.net/p/cmusphinx/code/trunk@10877 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ multisphinx/include/multisphinx/arc_buffer.h
@@ -75,8 +75,10 @@ typedef struct sarc_s {
      * unique language model state.
      */
     int16 lscr;
-    /** Index into an internal right context delta array, not for public use. */
-    int16 rc_idx;
+    /** Index into an internal right context delta array, not for
+     * public use.  (FIXME: suboptimal, it could be 16 bits if we
+     * gc'ed the rcdelta array properly) */
+    int32 rc_idx;
     /** Bitvector indicating active right contexts for this arc. */
     bitvec_t rc_bits[0];
 } sarc_t;
